core: Move SideName from ./metadata to ./side
Seems legit
